DB is standing basic requests working

+++ src/main/java/com/codecool/timecapsule/repository/UserRepository.java
@@ -3,5 +3,5 @@ package com.codecool.timecapsule.repository;
 import com.codecool.timecapsule.model.UserInformation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserInformation,Long> {
+public interface UserRepository extends JpaRepository<UserInformation,Integer> {
 }
